chore: move class file
